+++ javacup-core/src/main/java/io/cutalab/javacup/core/process/ProbeFailureKind.java
@@ -1,0 +1,11 @@
+package io.cutalab.javacup.core.process;
+
+public enum ProbeFailureKind {
+    NONE,
+    TIMEOUT,
+    PROCESS_NOT_FOUND,
+    ATTACH_FAILED,
+    JCMD_UNAVAILABLE,
+    INTERRUPTED,
+    UNKNOWN
+}